backend/requirements.txt
# ================= Python Version =================
# Use Python 3.10.11

# ================= Web Framework =================
Flask==3.0.0
flask-cors==4.0.0

# ================= Deep Learning =================
tensorflow==2.15.0
vit-keras==0.1.2

# ================= ML / Data =================
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# ================= Image Processing =================
opencv-python==4.8.1.78
Pillow==10.1.0

# ================= Visualization =================
matplotlib==3.8.2

# ================= Model Persistence =================
joblib==1.3.2

# ================= Chatbot =================
ollama

# ================= Utilities =================
requests==2.31.0